+++ build.js
@@ -45,8 +45,7 @@ async function buildAll() {
     'func',
     'load',
     'http',
-    'test',
-    'cloud_function',
+    'dev',
     'request',
     'server',
   ]) {